Used Long instead of Int

+++ app/src/main/kotlin/es/voghdev/playbattlegrounds/common/App.kt
@@ -70,9 +70,7 @@ class App : Application(), KodeinAware {
                     InsertMatchDBDataSource()
             )
         }
-        bind<IsAppExpired>() with singleton {
-            IsAppExpiredImpl(System.currentTimeMillis(), Limit.toLong())
-        }
+        bind<IsAppExpired>() with singleton { IsAppExpiredImpl(System.currentTimeMillis(), Limit) }
     }
 
     override fun onCreate() {
